backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
google-api-python-client
google-auth
python-dotenv
python-dateutil
